Converted all calculations and classes from double to float. Added the IsInteger method to the Literal class. Added the PolynomialSimplifier class. Added "unitstep" as a function name in the parser.
git-svn-id: file:///srv/svn/metaindu@314 aa2253e4-891b-cf47-b6be-b091cf79ccf8

+++ ArcsecantFunction.cs
@@ -29,7 +29,7 @@ namespace MetaphysicsIndustries.Solus
 
         protected override Literal InternalCall(VariableTable varTable, Literal[] args)
 		{
-            return new Literal(Math.Acos(1/args[0].Eval(varTable).Value));
+            return new Literal((float)Math.Acos(1 / args[0].Eval(varTable).Value));
         }
 
     }
