+++ Arctangent2Function.cs
@@ -11,9 +11,9 @@ namespace MetaphysicsIndustries.Solus
         {
         }
 
-        protected override double InternalCall(double arg0, double arg1)
+        protected override float InternalCall(float arg0, float arg1)
         {
-            return Math.Atan2(arg0, arg1);
+            return (float)Math.Atan2(arg0, arg1);
         }
 
         public override string DisplayName
