+++ CosecantFunction.cs
@@ -29,7 +29,7 @@ namespace MetaphysicsIndustries.Solus
 
         protected override Literal InternalCall(VariableTable varTable, Literal[] args)
 		{
-            return new Literal(1 / Math.Sin(args[0].Eval(varTable).Value));
+            return new Literal((float)(1 / Math.Sin(args[0].Eval(varTable).Value)));
         }
 
     }
